# (Performance) Change static readonly fields to const fields