--- a/Source/ACE.DatLoader/DatDatabase.cs
+++ b/Source/ACE.DatLoader/DatDatabase.cs
@@ -13,7 +13,7 @@ namespace ACE.DatLoader
     {
         private static readonly ILog log = LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
 
-        private static readonly uint DAT_HEADER_OFFSET = 0x140;
+        private const uint DAT_HEADER_OFFSET = 0x140;
 
         public string FilePath { get; }