--- a/Source/ACE.Database/WorldDatabaseWithEntityCache.cs
+++ b/Source/ACE.Database/WorldDatabaseWithEntityCache.cs
@@ -863,7 +863,7 @@ namespace ACE.Database
             }
         }
 
-        private static readonly float NormalizeEpsilon = 0.00001f;
+        private const float NormalizeEpsilon = 0.00001f;
 
         private void TreasureMaterialBase_Normalize(Dictionary<int, Dictionary<int, List<TreasureMaterialBase>>> materialBase)
         {