--- a/Source/ACE.Server/Entity/Cloak.cs
+++ b/Source/ACE.Server/Entity/Cloak.cs
@@ -77,7 +77,7 @@ namespace ACE.Server.Entity
             }
         }
 
-        private static readonly float TwoThirds = 2.0f / 3.0f;
+        private const float TwoThirds = 2.0f / 3.0f;
 
         /// <summary>
         /// Rolls for a chance at procing a cloak spell
@@ -197,7 +197,7 @@ namespace ACE.Server.Entity
         /// <summary>
         /// The amount of damage reduced by a cloak proced with PropertyInt.CloakWeaveProc=2
         /// </summary>
-        public static readonly int DamageReductionAmount = 200;
+        public const int DamageReductionAmount = 200;
 
         public static int GetDamageReductionAmount(WorldObject source)
         {