--- a/Source/ACE.Server/Entity/Mutations/MutationCache.cs
+++ b/Source/ACE.Server/Entity/Mutations/MutationCache.cs
@@ -512,7 +512,7 @@ namespace ACE.Server.Entity.Mutations
             return null;
         }
 
-        private static readonly string prefix = "ACE.Server.Entity.Mutations.";
+        private const string prefix = "ACE.Server.Entity.Mutations.";
 
         private static List<string> ReadScript(string filename)
         {