--- a/Source/ACE.Server/Factories/Tables/Cantrips/ArmorCantrips.cs
+++ b/Source/ACE.Server/Factories/Tables/Cantrips/ArmorCantrips.cs
@@ -95,7 +95,7 @@ namespace ACE.Server.Factories.Tables
             SpellId.CANTRIPPIERCINGBANE1,
         };
 
-        private static readonly int NumLevels = 4;
+        private const int NumLevels = 4;
 
         // original api
         public static readonly SpellId[][] Table = new SpellId[spells.Count][];