--- a/Source/ACE.Server/Factories/Tables/Cantrips/JewelryCantrips.cs
+++ b/Source/ACE.Server/Factories/Tables/Cantrips/JewelryCantrips.cs
@@ -73,7 +73,7 @@ namespace ACE.Server.Factories.Tables
             SpellId.CANTRIPPIERCINGWARD1,
         };
 
-        private static readonly int NumLevels = 4;
+        private const int NumLevels = 4;
 
         // original api
         public static readonly SpellId[][] Table = new SpellId[spells.Count][];