--- a/Source/ACE.Server/Factories/Tables/Cantrips/MeleeCantrips.cs
+++ b/Source/ACE.Server/Factories/Tables/Cantrips/MeleeCantrips.cs
@@ -29,7 +29,7 @@ namespace ACE.Server.Factories.Tables
             SpellId.CantripSneakAttackProwess1,
         };
 
-        private static readonly int NumLevels = 4;
+        private const int NumLevels = 4;
 
         // original api
         public static readonly SpellId[][] Table = new SpellId[spells.Count][];