--- a/Source/ACE.Server/Factories/Tables/Cantrips/MissileCantrips.cs
+++ b/Source/ACE.Server/Factories/Tables/Cantrips/MissileCantrips.cs
@@ -28,7 +28,7 @@ namespace ACE.Server.Factories.Tables
             SpellId.CantripSneakAttackProwess1,
         };
 
-        private static readonly int NumLevels = 4;
+        private const int NumLevels = 4;
 
         // original api
         public static readonly SpellId[][] Table = new SpellId[spells.Count][];