--- a/Source/ACE.Server/Factories/Tables/Cantrips/WandCantrips.cs
+++ b/Source/ACE.Server/Factories/Tables/Cantrips/WandCantrips.cs
@@ -32,7 +32,7 @@ namespace ACE.Server.Factories.Tables
             SpellId.CantripSpiritThirst1,
         };
 
-        private static readonly int NumLevels = 4;
+        private const int NumLevels = 4;
 
         // original api
         public static readonly SpellId[][] Table = new SpellId[spells.Count][];