--- a/Source/ACE.Server/Factories/Tables/Spells/ArmorSpells.cs
+++ b/Source/ACE.Server/Factories/Tables/Spells/ArmorSpells.cs
@@ -103,7 +103,7 @@ namespace ACE.Server.Factories.Tables
             SpellId.LightningBane1,
         };
 
-        private static readonly int NumTiers = 8;
+        private const int NumTiers = 8;
 
         // original api
         public static readonly SpellId[][] Table = new SpellId[spells.Count][];