--- a/Source/ACE.Server/Factories/Tables/Spells/GemSpells.cs
+++ b/Source/ACE.Server/Factories/Tables/Spells/GemSpells.cs
@@ -87,7 +87,7 @@ namespace ACE.Server.Factories.Tables
             SpellId.LightningProtectionSelf1,
         };
 
-        private static readonly int NumTiers = 8;
+        private const int NumTiers = 8;
 
         // original api
         public static readonly SpellId[][] GemCreatureSpellMatrix = new SpellId[NumTiers][];