--- a/Source/ACE.Server/Factories/Tables/Spells/MissileSpells.cs
+++ b/Source/ACE.Server/Factories/Tables/Spells/MissileSpells.cs
@@ -29,7 +29,7 @@ namespace ACE.Server.Factories.Tables
             SpellId.SneakAttackMasterySelf1,
         };
 
-        private static readonly int NumTiers = 8;
+        private const int NumTiers = 8;
 
         // original api
         public static readonly SpellId[][] Table = new SpellId[spells.Count][];