--- a/Source/ACE.Server/Factories/Tables/Spells/ScrollSpells.cs
+++ b/Source/ACE.Server/Factories/Tables/Spells/ScrollSpells.cs
@@ -346,9 +346,9 @@ namespace ACE.Server.Factories.Tables
             SpellId.CurseWeakness1,
         };
 
-        private static readonly int NumLevels = 7;
+        private const int NumLevels = 7;
 
-        private static readonly int MaxLevels = 8;
+        private const int MaxLevels = 8;
 
         private static readonly int NumSpells = creatureSpells.Count + lifeSpells.Count + itemSpells.Count + warSpells.Count + voidSpells.Count;