--- a/Source/ACE.Server/Network/GameEvent/Events/GameEventStartBarber.cs
+++ b/Source/ACE.Server/Network/GameEvent/Events/GameEventStartBarber.cs
@@ -2,8 +2,8 @@ namespace ACE.Server.Network.GameEvent.Events
 {
     public class GameEventStartBarber : GameEventMessage
     {
-        public static readonly uint EmpyreanMaleMotionDID   = 0x0900020E;
-        public static readonly uint EmpyreanFemaleMotionDID = 0x0900020D;
+        public const uint EmpyreanMaleMotionDID   = 0x0900020E;
+        public const uint EmpyreanFemaleMotionDID = 0x0900020D;
 
         public GameEventStartBarber(Session session)
             : base(GameEventType.StartBarber, GameMessageGroup.UIQueue, session, 68)