--- a/Source/ACE.Server/Network/Structure/RestrictionDB.cs
+++ b/Source/ACE.Server/Network/Structure/RestrictionDB.cs
@@ -79,11 +79,11 @@ namespace ACE.Server.Network.Structure
             writer.Write(restrictions.Table);
         }
 
-        private static readonly ushort headerNumBuckets = 768;  // this # of buckets was sent over the wire in retail header
+        private const ushort headerNumBuckets = 768;            // this # of buckets was sent over the wire in retail header
                                                                 // however, this value ends up being unused, and the "real" # of buckets originates
                                                                 // from a hardcoded value in g_bucketSizeArray in the client constant data
 
-        private static readonly ushort actualNumBuckets = 89;   // in RestrictionDB constructor in acclient,
+        private const ushort actualNumBuckets = 89;             // in RestrictionDB constructor in acclient,
                                                                 // client uses PHashTable for this (as opposed to the typical PackableHashTable)
 
                                                                 // which inits an IntrusiveHashTable with size 64