--- a/Source/ACE.Server/Physics/Animation/MovementParameters.cs
+++ b/Source/ACE.Server/Physics/Animation/MovementParameters.cs
@@ -42,12 +42,12 @@ namespace ACE.Server.Physics.Animation
         public HoldKey HoldKeyToApply;
         public int ActionStamp;
 
-        public static readonly float Default_DistanceToObject = 0.6f;
-        public static readonly float Default_FailDistance = float.MaxValue;
-        public static readonly float Default_MinDistance = 0.0f;
-        public static readonly float Default_Speed = 1.0f;
-        //public static readonly float Default_WalkRunThreshold = 15.0f;
-        public static readonly float Default_WalkRunThreshold = 1.0f;
+        public const float Default_DistanceToObject = 0.6f;
+        public const float Default_FailDistance = float.MaxValue;
+        public const float Default_MinDistance = 0.0f;
+        public const float Default_Speed = 1.0f;
+        //public const float Default_WalkRunThreshold = 15.0f;
+        public const float Default_WalkRunThreshold = 1.0f;
 
         public MovementParameters()
         {