--- a/Source/ACE.Server/Physics/Sphere.cs
+++ b/Source/ACE.Server/Physics/Sphere.cs
@@ -60,8 +60,8 @@ namespace ACE.Server.Physics
             Radius = sphere.Radius;
         }
 
-        public static readonly float ThresholdMed = 1.0f / 3.0f;
-        public static readonly float ThresholdHigh = 2.0f / 3.0f;
+        public const float ThresholdMed = 1.0f / 3.0f;
+        public const float ThresholdHigh = 2.0f / 3.0f;
 
         public static Quadrant Attack(Position targetPos, float targetRadius, float targetHeight, Position attackPos, Vector2 left, Vector2 right, float attackRadius, float attackHeight)
         {