… Source/ACE.Server/Physics/PhysicsGlobals.cs
… Source/ACE.Server/WorldObjects/Creature_Melee.cs
… Source/ACE.Server/WorldObjects/Creature_Missile.cs
… Source/ACE.Server/WorldObjects/Healer.cs
… Source/ACE.Server/WorldObjects/Managers/ConfirmationManager.cs
… Source/ACE.Server/WorldObjects/Monster_Awareness.cs
… Source/ACE.Server/WorldObjects/Monster_Magic.cs
… Source/ACE.Server/WorldObjects/Monster_Missile.cs
… Source/ACE.Server/WorldObjects/Monster_Navigation.cs
… Source/ACE.Server/WorldObjects/Pet.cs
… Source/ACE.Server/WorldObjects/Player.cs
… Source/ACE.Server/WorldObjects/Player_Combat.cs
… Source/ACE.Server/WorldObjects/Player_Commerce.cs
… Source/ACE.Server/WorldObjects/Player_Database.cs
… Source/ACE.Server/WorldObjects/Player_Death.cs
… Source/ACE.Server/WorldObjects/Player_Location.cs
… Source/ACE.Server/WorldObjects/Player_Magic.cs
… Source/ACE.Server/WorldObjects/Player_Melee.cs
… Source/ACE.Server/WorldObjects/Portal.cs
… Source/ACE.Server/WorldObjects/SkillFormula.cs
… Source/ACE.Server/WorldObjects/Vendor.cs
… Source/ACE.Server/WorldObjects/WorldObject.cs
… Source/ACE.Server/WorldObjects/WorldObject_Magic.cs
… Source/ACE.Server/WorldObjects/WorldObject_Weapon.cs