This will at least somewhat improve performance, as confirmed through microbenchmark testing.

const fields are replaced with literal values in the Assembly's IL code, while static readonly fields may be optimized by the JIT, and is not guaranteed to happen. const fields may also lend themselves to stronger optimizations at runtime.

Method is [LandDefs.AdjustToOutside](https://github.com/ACEmulator/ACE/blob/89e023c9d3dc2600a720a59229d1edeb2b5b0402/Source/ACE.Server/Physics/Common/LandDefs.cs#L125-L147).
Baseline (static readonly):
```
| Method          | Mean     | Error    | StdDev   | Allocated |
|---------------- |---------:|---------:|---------:|----------:|
| AdjustToOutside | 15.90 ns | 0.300 ns | 0.308 ns |         - |
```
With [constants](https://github.com/Flaggs-ACE-Contributions/ACE/commit/7709a5820b385adfa032a64f01bead6b6418c9ca#diff-a1f4bef19969e8833e3f787e6779be4285d57d14a9ea30b1f6b39d92d20b0a41L86-R110):
```
| Method          | Mean     | Error    | StdDev   | Allocated |
|---------------- |---------:|---------:|---------:|----------:|
| AdjustToOutside | 13.36 ns | 0.243 ns | 0.227 ns |         - |
```

This is a 15% improvement for LandDefs.AdjustToOutside. Additionally, I measured through some [custom profiler code](https://github.com/ACRealms/ACRealms.WorldServer/commit/9c58704106b30917d1ab6b680292060f14d51eee) that LandDefs.AdjustToOutside is called [very frequently](https://gist.github.com/FlaggAC/fb862223306e454f81279883c95e0d8b). So I would estimate this to be a solid improvement.